requirements.txt
# --- Web/API ---
fastapi==0.95.2          # pydantic v1 uyumu (schemas.py bu sürüme göre)
uvicorn[standard]==0.23.2

# --- DB/ORM ---
SQLAlchemy>=2.0.35
psycopg[binary]>=3.2.0  # Postgres kullanacaksan; SQLite için şart değil

# --- HTTP/Async ---
httpx==0.27.0
requests==2.31.0
redis==5.0.4
aiohttp==3.9.3

# --- LLM ---
openai==1.43.0           # llm_client.py ile uyumlu (from openai import OpenAI)
google-generativeai==0.8.0  # Google Gemini API desteği
groq==0.9.0              # Groq API desteği (ücretsiz, çok hızlı)

# --- Yardımcılar ---
python-dotenv==1.0.1
cryptography==42.0.5

# --- Monitoring ---
prometheus-client==0.20.0  # Prometheus metrics exporter

# --- Testing ---
pytest==8.2.1
